Data set for aggregate level patient data in wide format

This data set contains summaries of simulated patient covariate and outcome values.

Format

A data list including the variables needed for a outstandR basic example. The variables are as follows:

mean.X1

Numeric value

mean.X2

Numeric value

mean.X3

Numeric value

mean.X4

Numeric value

sd.X1

Numeric standard deviation value

sd.X2

Numeric standard deviation value

sd.X3

Numeric standard deviation value

sd.X4

Numeric standard deviation value

y.B.sum

Integer number of patients with event on B treatment

y.B.bar

Numeric proportion of patients on B treatment

N.B

Integer Total number of patients on B treatment

y.C.sum

Integer number of patients with event on C treatment

y.C.bar

Numeric proportion of patients on C treatment

N.C

Integer total number of patients on C treatment

Source

Simulated data

References

Remiro‐Azócar A, Heath A, Baio G (2022)