Calculate individual-level patient data statistics

Separate methods for each approach MAIC, STC, G-computation via MLE or Bayesian inference.

Usage

IPD_stats(strategy, ipd, ald, scale, ...)

# Default S3 method
IPD_stats(...)

# S3 method for class 'mim'
IPD_stats(strategy, ipd, ald, scale, ...)

# S3 method for class 'stc'
IPD_stats(strategy, ipd, ald, scale, ...)

# S3 method for class 'maic'
IPD_stats(strategy, ipd, ald, scale, ...)

# S3 method for class 'gcomp_ml'
IPD_stats(strategy, ipd, ald, scale, ...)

# S3 method for class 'gcomp_stan'
IPD_stats(strategy, ipd, ald, scale, ...)

Arguments

strategy

A list corresponding to different approaches

ipd

Individual-level data

ald

Aggregate-level data

...

Additional arguments

Value

Mean and variance values

Multiple imputation marginalisation

Using Stan, compute marginal relative treatment effect for A vs C for each MCMC sample by transforming from probability to linear predictor scale. Approximate by using imputation and combining estimates using Rubin's rules, in contrast to IPD_stats.gcomp_stan().

Simulated treatment comparison statistics

IPD from the AC trial are used to fit a regression model describing the observed outcomes \(y\) in terms of the relevant baseline characteristics \(x\) and the treatment variable \(z\).

Matching-adjusted indirect comparison statistics

Marginal A vs C treatment effect estimates using bootstrapping sampling.

G-computation maximum likelihood statistics

Compute a non-parametric bootstrap with default \(R=1000\) resamples.

G-computation Bayesian statistics

Using Stan, compute marginal log-odds ratio for A vs C for each MCMC sample by transforming from probability to linear predictor scale.