Bayesian G-computation using Stan

Calculate draws of binary responses from posterior predictive distribution from the Bayesian G-computation method using Hamiltonian Monte Carlo.

Usage

calc_gcomp_bayes(strategy, analysis_params, ...)

Arguments

strategy

A list specifying the model strategy, including:

formula: A linear regression formula object.

family: A family object specifying the distribution and link function (e.g., binomial).

iter: Number of iterations for the MCMC sampling.

warmup: Number of warmup iterations for the MCMC sampling.

chains: Number of MCMC chains.

analysis_params

List of analysis parameters. Must contain ipd and ald.

...

Additional arguments passed to rstanarm::stan_glm().

Value

A list containing:

means: A list containing:

A: Posterior means for comparator treatment group "A".

C: Posterior means for reference treatment group "C".

model: A list containing the fit object (from stan_glm), rho, N, and stan_args.

Examples

strategy <- list(
  formula = y ~ trt:X1,
  family = binomial(),
  rho = NA,
  N = 1000L,
  marginal_distns = NA,
  marginal_params = NA,
  trt_var = "trt",
  iter = 2000,
  warmup = 500,
  chains = 4)

ipd <- data.frame(
   trt = sample(c("A", "C"), size = 100, replace = TRUE),
   X1 = rnorm(100, 1, 1),
   y = sample(c(1,0), size = 100, prob = c(0.7, 0.3), replace = TRUE))

ald <- data.frame(
  trt = c(NA, NA, "B", "C", "B", "C"),
  variable = c("X1", "X1", "y", "y", NA, NA),
  statistic = c("mean", "sd", "sum", "sum", "N", "N"),
  value = c(0.5, 0.1, 10, 12, 20, 25))

calc_gcomp_bayes(
  strategy,
  analysis_params = list(
    ipd = ipd, ald = ald, 
    ref_trt = "C",
    ipd_comp = "A"))
#> $means
#> $means$A
#>    [1] 0.675 0.553 0.606 0.613 0.592 0.782 0.542 0.510 0.522 0.546 0.601 0.580
#>   [13] 0.650 0.628 0.607 0.557 0.618 0.657 0.647 0.673 0.685 0.670 0.676 0.652
#>   [25] 0.571 0.567 0.550 0.669 0.655 0.654 0.424 0.509 0.738 0.684 0.637 0.666
#>   [37] 0.648 0.679 0.576 0.614 0.588 0.603 0.625 0.633 0.617 0.668 0.650 0.691
#>   [49] 0.584 0.668 0.607 0.707 0.716 0.645 0.568 0.617 0.578 0.557 0.690 0.732
#>   [61] 0.666 0.707 0.640 0.654 0.578 0.613 0.659 0.697 0.758 0.704 0.727 0.667
#>   [73] 0.625 0.585 0.621 0.686 0.690 0.566 0.615 0.680 0.644 0.688 0.732 0.648
#>   [85] 0.610 0.662 0.603 0.522 0.627 0.633 0.670 0.605 0.569 0.751 0.590 0.679
#>   [97] 0.691 0.692 0.673 0.711 0.693 0.696 0.718 0.665 0.590 0.649 0.597 0.636
#>  [109] 0.550 0.698 0.543 0.535 0.678 0.655 0.546 0.635 0.661 0.650 0.562 0.599
#>  [121] 0.696 0.654 0.628 0.600 0.699 0.593 0.651 0.654 0.650 0.683 0.734 0.650
#>  [133] 0.614 0.407 0.752 0.666 0.691 0.647 0.698 0.584 0.643 0.691 0.626 0.592
#>  [145] 0.699 0.662 0.626 0.665 0.644 0.606 0.622 0.631 0.662 0.646 0.585 0.631
#>  [157] 0.733 0.713 0.620 0.758 0.709 0.593 0.688 0.635 0.627 0.650 0.720 0.631
#>  [169] 0.633 0.625 0.661 0.701 0.684 0.709 0.612 0.637 0.626 0.667 0.569 0.675
#>  [181] 0.592 0.579 0.600 0.731 0.474 0.581 0.643 0.642 0.673 0.691 0.666 0.683
#>  [193] 0.623 0.675 0.656 0.639 0.665 0.601 0.595 0.664 0.677 0.766 0.733 0.579
#>  [205] 0.588 0.545 0.673 0.652 0.544 0.552 0.517 0.682 0.698 0.700 0.667 0.647
#>  [217] 0.698 0.621 0.729 0.615 0.634 0.689 0.615 0.673 0.605 0.675 0.698 0.726
#>  [229] 0.661 0.675 0.663 0.722 0.607 0.719 0.587 0.566 0.667 0.592 0.549 0.643
#>  [241] 0.650 0.688 0.654 0.646 0.633 0.640 0.551 0.547 0.728 0.702 0.711 0.691
#>  [253] 0.698 0.552 0.514 0.446 0.563 0.519 0.540 0.485 0.696 0.634 0.708 0.694
#>  [265] 0.619 0.541 0.752 0.759 0.706 0.561 0.785 0.704 0.718 0.676 0.590 0.650
#>  [277] 0.781 0.666 0.668 0.675 0.681 0.739 0.616 0.639 0.633 0.618 0.660 0.672
#>  [289] 0.676 0.679 0.725 0.530 0.716 0.663 0.675 0.599 0.563 0.630 0.591 0.706
#>  [301] 0.553 0.596 0.698 0.599 0.595 0.706 0.618 0.638 0.597 0.709 0.573 0.603
#>  [313] 0.718 0.640 0.571 0.670 0.566 0.594 0.586 0.556 0.675 0.642 0.621 0.638
#>  [325] 0.626 0.664 0.668 0.696 0.765 0.629 0.668 0.572 0.686 0.623 0.741 0.607
#>  [337] 0.715 0.680 0.696 0.601 0.608 0.740 0.657 0.627 0.594 0.610 0.660 0.614
#>  [349] 0.629 0.594 0.582 0.698 0.685 0.645 0.652 0.514 0.696 0.698 0.708 0.653
#>  [361] 0.619 0.649 0.668 0.614 0.695 0.705 0.664 0.717 0.687 0.734 0.677 0.578
#>  [373] 0.589 0.570 0.633 0.665 0.664 0.525 0.687 0.680 0.685 0.572 0.640 0.682
#>  [385] 0.618 0.597 0.605 0.659 0.612 0.652 0.640 0.697 0.734 0.620 0.698 0.651
#>  [397] 0.614 0.634 0.607 0.760 0.547 0.423 0.675 0.601 0.715 0.623 0.732 0.625
#>  [409] 0.601 0.559 0.647 0.591 0.604 0.690 0.642 0.602 0.605 0.651 0.691 0.588
#>  [421] 0.710 0.550 0.689 0.671 0.694 0.515 0.742 0.633 0.697 0.536 0.739 0.632
#>  [433] 0.648 0.668 0.635 0.719 0.679 0.649 0.527 0.709 0.657 0.669 0.697 0.692
#>  [445] 0.690 0.692 0.583 0.741 0.759 0.643 0.550 0.546 0.574 0.573 0.528 0.562
#>  [457] 0.620 0.738 0.774 0.544 0.587 0.560 0.531 0.571 0.529 0.516 0.553 0.638
#>  [469] 0.636 0.679 0.661 0.688 0.628 0.679 0.723 0.637 0.670 0.637 0.646 0.709
#>  [481] 0.595 0.641 0.496 0.604 0.594 0.708 0.535 0.548 0.575 0.564 0.585 0.689
#>  [493] 0.675 0.656 0.663 0.676 0.679 0.572 0.595 0.601 0.712 0.651 0.740 0.797
#>  [505] 0.619 0.557 0.576 0.634 0.660 0.635 0.663 0.576 0.668 0.586 0.605 0.614
#>  [517] 0.737 0.605 0.607 0.634 0.614 0.693 0.715 0.549 0.714 0.686 0.642 0.611
#>  [529] 0.603 0.666 0.657 0.567 0.678 0.643 0.652 0.593 0.692 0.567 0.697 0.587
#>  [541] 0.659 0.674 0.590 0.651 0.563 0.584 0.607 0.688 0.709 0.571 0.594 0.635
#>  [553] 0.637 0.584 0.655 0.638 0.765 0.680 0.674 0.578 0.677 0.666 0.662 0.582
#>  [565] 0.543 0.715 0.562 0.687 0.671 0.629 0.527 0.648 0.543 0.669 0.735 0.746
#>  [577] 0.581 0.501 0.577 0.614 0.559 0.674 0.526 0.530 0.629 0.712 0.711 0.638
#>  [589] 0.582 0.647 0.676 0.651 0.690 0.737 0.727 0.683 0.635 0.632 0.656 0.633
#>  [601] 0.719 0.718 0.691 0.606 0.664 0.637 0.622 0.556 0.737 0.760 0.681 0.666
#>  [613] 0.638 0.725 0.737 0.773 0.491 0.534 0.723 0.781 0.743 0.673 0.570 0.561
#>  [625] 0.595 0.574 0.667 0.601 0.679 0.662 0.633 0.595 0.684 0.573 0.728 0.686
#>  [637] 0.722 0.564 0.571 0.711 0.625 0.714 0.770 0.715 0.739 0.658 0.650 0.612
#>  [649] 0.680 0.654 0.760 0.573 0.740 0.617 0.573 0.654 0.612 0.677 0.631 0.620
#>  [661] 0.580 0.594 0.655 0.611 0.659 0.744 0.665 0.667 0.655 0.728 0.578 0.628
#>  [673] 0.637 0.669 0.581 0.629 0.602 0.609 0.637 0.569 0.523 0.692 0.646 0.720
#>  [685] 0.732 0.552 0.543 0.551 0.652 0.661 0.596 0.574 0.676 0.611 0.662 0.692
#>  [697] 0.566 0.609 0.609 0.711 0.574 0.543 0.577 0.610 0.716 0.618 0.619 0.652
#>  [709] 0.623 0.601 0.685 0.570 0.673 0.642 0.651 0.610 0.619 0.715 0.588 0.665
#>  [721] 0.611 0.621 0.569 0.526 0.689 0.623 0.610 0.716 0.521 0.697 0.542 0.541
#>  [733] 0.627 0.662 0.675 0.688 0.645 0.633 0.628 0.701 0.601 0.634 0.634 0.571
#>  [745] 0.625 0.575 0.681 0.590 0.625 0.771 0.632 0.665 0.656 0.610 0.798 0.598
#>  [757] 0.637 0.757 0.719 0.559 0.490 0.459 0.792 0.790 0.737 0.689 0.664 0.608
#>  [769] 0.640 0.640 0.625 0.700 0.590 0.580 0.643 0.583 0.562 0.573 0.709 0.637
#>  [781] 0.646 0.720 0.569 0.555 0.713 0.669 0.612 0.693 0.723 0.737 0.647 0.608
#>  [793] 0.553 0.744 0.610 0.695 0.624 0.647 0.616 0.655 0.557 0.634 0.655 0.622
#>  [805] 0.616 0.773 0.709 0.676 0.605 0.614 0.684 0.635 0.652 0.637 0.646 0.649
#>  [817] 0.632 0.591 0.530 0.679 0.738 0.723 0.707 0.736 0.729 0.640 0.565 0.683
#>  [829] 0.575 0.619 0.738 0.579 0.639 0.735 0.789 0.792 0.732 0.565 0.592 0.603
#>  [841] 0.599 0.671 0.628 0.646 0.594 0.652 0.701 0.660 0.761 0.638 0.677 0.704
#>  [853] 0.614 0.598 0.583 0.612 0.663 0.640 0.636 0.601 0.698 0.676 0.671 0.722
#>  [865] 0.603 0.600 0.637 0.750 0.653 0.733 0.666 0.688 0.668 0.657 0.627 0.662
#>  [877] 0.682 0.628 0.669 0.681 0.607 0.645 0.666 0.671 0.643 0.607 0.663 0.623
#>  [889] 0.617 0.589 0.591 0.609 0.637 0.658 0.610 0.635 0.643 0.583 0.684 0.692
#>  [901] 0.728 0.665 0.617 0.591 0.628 0.606 0.731 0.765 0.539 0.515 0.664 0.704
#>  [913] 0.695 0.701 0.679 0.666 0.590 0.659 0.648 0.622 0.666 0.598 0.581 0.583
#>  [925] 0.650 0.655 0.680 0.620 0.638 0.626 0.587 0.652 0.569 0.609 0.636 0.693
#>  [937] 0.612 0.696 0.691 0.651 0.620 0.555 0.592 0.603 0.744 0.580 0.701 0.606
#>  [949] 0.671 0.701 0.537 0.663 0.592 0.669 0.651 0.635 0.660 0.628 0.611 0.709
#>  [961] 0.587 0.549 0.710 0.735 0.661 0.598 0.564 0.674 0.673 0.685 0.639 0.642
#>  [973] 0.656 0.596 0.627 0.619 0.636 0.634 0.597 0.554 0.597 0.623 0.693 0.649
#>  [985] 0.636 0.687 0.725 0.650 0.631 0.597 0.638 0.695 0.585 0.662 0.693 0.615
#>  [997] 0.514 0.634 0.676 0.663 0.638 0.724 0.721 0.691 0.629 0.643 0.678 0.677
#> [1009] 0.658 0.689 0.597 0.609 0.656 0.708 0.542 0.669 0.645 0.626 0.610 0.637
#> [1021] 0.608 0.742 0.551 0.582 0.703 0.749 0.640 0.643 0.658 0.683 0.604 0.615
#> [1033] 0.614 0.703 0.728 0.654 0.652 0.617 0.600 0.573 0.680 0.625 0.674 0.670
#> [1045] 0.685 0.651 0.671 0.653 0.652 0.657 0.657 0.611 0.626 0.673 0.651 0.637
#> [1057] 0.673 0.633 0.587 0.538 0.532 0.583 0.704 0.687 0.721 0.677 0.687 0.660
#> [1069] 0.697 0.700 0.702 0.739 0.716 0.576 0.742 0.732 0.772 0.682 0.641 0.637
#> [1081] 0.621 0.640 0.679 0.627 0.609 0.628 0.674 0.582 0.588 0.585 0.572 0.583
#> [1093] 0.679 0.633 0.634 0.662 0.647 0.617 0.631 0.603 0.528 0.542 0.613 0.640
#> [1105] 0.724 0.660 0.612 0.591 0.617 0.669 0.674 0.654 0.671 0.652 0.587 0.725
#> [1117] 0.651 0.691 0.702 0.673 0.468 0.507 0.459 0.661 0.680 0.565 0.701 0.725
#> [1129] 0.639 0.624 0.630 0.615 0.615 0.786 0.677 0.723 0.630 0.680 0.634 0.673
#> [1141] 0.592 0.509 0.587 0.706 0.593 0.632 0.646 0.659 0.665 0.559 0.697 0.658
#> [1153] 0.602 0.579 0.630 0.655 0.585 0.713 0.608 0.651 0.608 0.665 0.642 0.663
#> [1165] 0.671 0.679 0.703 0.637 0.634 0.612 0.624 0.683 0.657 0.660 0.659 0.612
#> [1177] 0.639 0.758 0.676 0.650 0.628 0.590 0.702 0.525 0.694 0.690 0.626 0.728
#> [1189] 0.596 0.655 0.638 0.598 0.716 0.717 0.789 0.470 0.671 0.596 0.645 0.700
#> [1201] 0.679 0.596 0.636 0.700 0.670 0.700 0.654 0.649 0.598 0.649 0.719 0.668
#> [1213] 0.664 0.690 0.640 0.657 0.682 0.645 0.659 0.606 0.650 0.652 0.667 0.704
#> [1225] 0.607 0.592 0.659 0.626 0.639 0.629 0.606 0.591 0.647 0.756 0.756 0.750
#> [1237] 0.618 0.657 0.593 0.691 0.719 0.641 0.650 0.540 0.713 0.584 0.701 0.615
#> [1249] 0.685 0.643 0.640 0.694 0.640 0.698 0.648 0.579 0.606 0.590 0.606 0.610
#> [1261] 0.603 0.656 0.645 0.693 0.650 0.659 0.541 0.689 0.596 0.630 0.652 0.672
#> [1273] 0.597 0.620 0.588 0.653 0.619 0.616 0.573 0.672 0.645 0.765 0.719 0.762
#> [1285] 0.642 0.577 0.690 0.610 0.615 0.661 0.533 0.613 0.669 0.680 0.670 0.615
#> [1297] 0.656 0.644 0.714 0.646 0.639 0.653 0.665 0.596 0.609 0.667 0.595 0.707
#> [1309] 0.683 0.682 0.636 0.575 0.599 0.681 0.608 0.653 0.541 0.564 0.765 0.529
#> [1321] 0.612 0.627 0.588 0.638 0.680 0.603 0.682 0.581 0.674 0.562 0.571 0.756
#> [1333] 0.649 0.614 0.600 0.691 0.569 0.674 0.645 0.666 0.660 0.747 0.738 0.705
#> [1345] 0.664 0.665 0.653 0.674 0.650 0.551 0.571 0.602 0.650 0.688 0.670 0.617
#> [1357] 0.687 0.698 0.671 0.672 0.694 0.723 0.636 0.700 0.623 0.584 0.638 0.602
#> [1369] 0.548 0.786 0.643 0.597 0.680 0.599 0.622 0.664 0.657 0.686 0.725 0.698
#> [1381] 0.681 0.701 0.519 0.728 0.649 0.715 0.670 0.646 0.590 0.631 0.668 0.499
#> [1393] 0.752 0.676 0.711 0.695 0.704 0.609 0.653 0.691 0.603 0.566 0.665 0.688
#> [1405] 0.627 0.767 0.558 0.583 0.537 0.492 0.550 0.723 0.414 0.858 0.536 0.562
#> [1417] 0.594 0.686 0.670 0.552 0.515 0.551 0.545 0.576 0.684 0.656 0.712 0.700
#> [1429] 0.663 0.646 0.668 0.563 0.630 0.641 0.695 0.609 0.739 0.645 0.680 0.698
#> [1441] 0.651 0.572 0.666 0.643 0.518 0.569 0.719 0.729 0.554 0.705 0.717 0.642
#> [1453] 0.686 0.655 0.723 0.669 0.649 0.693 0.565 0.577 0.661 0.677 0.689 0.648
#> [1465] 0.691 0.666 0.567 0.545 0.745 0.676 0.710 0.605 0.738 0.762 0.519 0.637
#> [1477] 0.668 0.680 0.645 0.748 0.558 0.601 0.691 0.654 0.679 0.555 0.561 0.668
#> [1489] 0.667 0.719 0.567 0.573 0.593 0.577 0.696 0.653 0.693 0.752 0.678 0.589
#> [1501] 0.628 0.629 0.650 0.698 0.680 0.617 0.671 0.560 0.612 0.794 0.509 0.609
#> [1513] 0.678 0.566 0.713 0.715 0.622 0.633 0.709 0.560 0.678 0.604 0.636 0.631
#> [1525] 0.617 0.786 0.719 0.713 0.594 0.687 0.616 0.587 0.646 0.628 0.702 0.678
#> [1537] 0.717 0.564 0.715 0.621 0.587 0.789 0.715 0.600 0.664 0.655 0.630 0.723
#> [1549] 0.618 0.652 0.588 0.635 0.695 0.703 0.690 0.654 0.788 0.758 0.783 0.745
#> [1561] 0.744 0.753 0.706 0.751 0.472 0.482 0.744 0.680 0.656 0.776 0.645 0.734
#> [1573] 0.619 0.534 0.596 0.748 0.532 0.655 0.646 0.584 0.650 0.543 0.615 0.597
#> [1585] 0.647 0.664 0.698 0.552 0.621 0.672 0.758 0.643 0.648 0.523 0.523 0.596
#> [1597] 0.546 0.726 0.627 0.611 0.588 0.619 0.666 0.718 0.628 0.520 0.718 0.689
#> [1609] 0.670 0.646 0.624 0.575 0.575 0.675 0.700 0.555 0.692 0.651 0.615 0.687
#> [1621] 0.685 0.656 0.638 0.682 0.568 0.679 0.695 0.688 0.684 0.740 0.604 0.568
#> [1633] 0.627 0.631 0.598 0.646 0.571 0.639 0.696 0.686 0.634 0.620 0.666 0.607
#> [1645] 0.585 0.707 0.739 0.747 0.553 0.714 0.627 0.571 0.623 0.715 0.661 0.669
#> [1657] 0.617 0.665 0.706 0.734 0.706 0.564 0.695 0.738 0.744 0.620 0.570 0.698
#> [1669] 0.659 0.633 0.551 0.720 0.572 0.743 0.580 0.617 0.615 0.632 0.657 0.740
#> [1681] 0.549 0.492 0.577 0.734 0.647 0.632 0.678 0.656 0.671 0.680 0.586 0.637
#> [1693] 0.600 0.568 0.707 0.654 0.625 0.604 0.617 0.694 0.698 0.644 0.744 0.737
#> [1705] 0.562 0.501 0.615 0.557 0.730 0.762 0.677 0.705 0.671 0.697 0.736 0.580
#> [1717] 0.765 0.702 0.643 0.662 0.649 0.623 0.623 0.657 0.701 0.566 0.601 0.719
#> [1729] 0.727 0.621 0.664 0.636 0.671 0.722 0.760 0.659 0.659 0.581 0.702 0.559
#> [1741] 0.608 0.638 0.706 0.626 0.594 0.642 0.670 0.734 0.746 0.682 0.656 0.627
#> [1753] 0.646 0.584 0.579 0.733 0.744 0.586 0.638 0.636 0.761 0.760 0.745 0.536
#> [1765] 0.629 0.663 0.624 0.617 0.650 0.639 0.559 0.695 0.715 0.758 0.497 0.713
#> [1777] 0.798 0.494 0.662 0.650 0.683 0.664 0.573 0.646 0.693 0.549 0.554 0.688
#> [1789] 0.648 0.753 0.659 0.647 0.640 0.609 0.677 0.637 0.660 0.644 0.638 0.625
#> [1801] 0.650 0.665 0.645 0.667 0.698 0.665 0.662 0.653 0.806 0.495 0.755 0.565
#> [1813] 0.647 0.634 0.583 0.629 0.705 0.643 0.550 0.637 0.712 0.665 0.679 0.660
#> [1825] 0.655 0.670 0.686 0.671 0.586 0.695 0.662 0.612 0.656 0.657 0.658 0.631
#> [1837] 0.726 0.783 0.717 0.552 0.636 0.705 0.577 0.580 0.631 0.639 0.633 0.639
#> [1849] 0.624 0.604 0.575 0.709 0.661 0.672 0.655 0.579 0.571 0.629 0.676 0.608
#> [1861] 0.622 0.690 0.689 0.682 0.582 0.594 0.674 0.728 0.631 0.643 0.649 0.638
#> [1873] 0.657 0.656 0.703 0.665 0.596 0.573 0.666 0.677 0.653 0.640 0.653 0.679
#> [1885] 0.628 0.666 0.663 0.616 0.663 0.688 0.668 0.620 0.606 0.701 0.574 0.665
#> [1897] 0.588 0.724 0.791 0.765 0.736 0.548 0.704 0.441 0.771 0.678 0.663 0.619
#> [1909] 0.738 0.469 0.643 0.604 0.654 0.657 0.594 0.587 0.670 0.632 0.637 0.694
#> [1921] 0.665 0.591 0.596 0.575 0.489 0.538 0.667 0.689 0.663 0.648 0.658 0.635
#> [1933] 0.664 0.689 0.700 0.683 0.632 0.604 0.610 0.684 0.621 0.706 0.719 0.715
#> [1945] 0.746 0.652 0.625 0.655 0.610 0.624 0.695 0.580 0.613 0.595 0.751 0.584
#> [1957] 0.618 0.576 0.551 0.757 0.750 0.743 0.593 0.577 0.724 0.759 0.689 0.573
#> [1969] 0.610 0.627 0.564 0.631 0.728 0.598 0.582 0.665 0.707 0.663 0.721 0.570
#> [1981] 0.515 0.754 0.550 0.671 0.560 0.600 0.636 0.686 0.669 0.706 0.701 0.511
#> [1993] 0.550 0.490 0.748 0.820 0.452 0.400 0.475 0.690
#> 
#> $means$C
#>    [1] 0.638 0.530 0.582 0.661 0.633 0.708 0.606 0.523 0.554 0.591 0.649 0.598
#>   [13] 0.673 0.640 0.584 0.535 0.558 0.670 0.634 0.590 0.632 0.640 0.655 0.654
#>   [25] 0.517 0.504 0.522 0.644 0.626 0.633 0.488 0.476 0.677 0.622 0.564 0.594
#>   [37] 0.533 0.616 0.599 0.601 0.617 0.577 0.577 0.574 0.673 0.652 0.669 0.696
#>   [49] 0.498 0.552 0.567 0.720 0.690 0.674 0.580 0.571 0.546 0.528 0.589 0.662
#>   [61] 0.697 0.596 0.705 0.642 0.542 0.554 0.658 0.709 0.761 0.716 0.651 0.661
#>   [73] 0.543 0.575 0.587 0.658 0.644 0.576 0.623 0.665 0.602 0.679 0.655 0.649
#>   [85] 0.581 0.687 0.584 0.505 0.625 0.614 0.663 0.589 0.559 0.719 0.561 0.651
#>   [97] 0.676 0.664 0.632 0.737 0.698 0.721 0.735 0.638 0.602 0.677 0.597 0.639
#>  [109] 0.554 0.677 0.518 0.510 0.537 0.624 0.567 0.585 0.627 0.637 0.531 0.586
#>  [121] 0.663 0.631 0.605 0.598 0.660 0.560 0.571 0.695 0.662 0.574 0.773 0.644
#>  [133] 0.569 0.430 0.715 0.631 0.619 0.639 0.661 0.617 0.589 0.729 0.588 0.574
#>  [145] 0.650 0.600 0.614 0.602 0.638 0.530 0.532 0.539 0.637 0.606 0.571 0.601
#>  [157] 0.669 0.636 0.616 0.680 0.648 0.596 0.636 0.661 0.663 0.653 0.653 0.660
#>  [169] 0.650 0.603 0.599 0.714 0.620 0.652 0.614 0.615 0.583 0.600 0.604 0.604
#>  [181] 0.562 0.497 0.661 0.727 0.476 0.645 0.631 0.566 0.652 0.744 0.668 0.641
#>  [193] 0.594 0.689 0.590 0.649 0.609 0.594 0.521 0.656 0.613 0.714 0.707 0.588
#>  [205] 0.577 0.539 0.582 0.588 0.528 0.563 0.555 0.623 0.694 0.675 0.660 0.622
#>  [217] 0.711 0.658 0.720 0.604 0.595 0.638 0.605 0.661 0.584 0.598 0.624 0.719
#>  [229] 0.642 0.685 0.584 0.720 0.558 0.680 0.608 0.477 0.640 0.602 0.503 0.547
#>  [241] 0.631 0.655 0.680 0.602 0.582 0.608 0.588 0.503 0.760 0.638 0.682 0.725
#>  [253] 0.665 0.534 0.519 0.466 0.527 0.557 0.487 0.467 0.678 0.570 0.628 0.664
#>  [265] 0.616 0.542 0.696 0.716 0.688 0.468 0.702 0.700 0.675 0.680 0.501 0.656
#>  [277] 0.697 0.708 0.567 0.616 0.668 0.667 0.676 0.694 0.526 0.568 0.649 0.634
#>  [289] 0.664 0.639 0.683 0.470 0.704 0.624 0.634 0.553 0.604 0.599 0.616 0.629
#>  [301] 0.561 0.586 0.670 0.547 0.552 0.609 0.601 0.651 0.598 0.715 0.517 0.633
#>  [313] 0.608 0.574 0.525 0.612 0.546 0.556 0.615 0.569 0.619 0.528 0.536 0.611
#>  [325] 0.602 0.629 0.674 0.712 0.709 0.618 0.649 0.581 0.672 0.611 0.701 0.566
#>  [337] 0.651 0.659 0.653 0.561 0.561 0.711 0.632 0.602 0.615 0.589 0.571 0.580
#>  [349] 0.599 0.583 0.574 0.666 0.690 0.668 0.653 0.491 0.648 0.694 0.676 0.702
#>  [361] 0.521 0.567 0.595 0.640 0.686 0.634 0.630 0.647 0.645 0.650 0.658 0.584
#>  [373] 0.608 0.655 0.669 0.594 0.589 0.523 0.606 0.684 0.654 0.546 0.634 0.678
#>  [385] 0.626 0.629 0.605 0.645 0.629 0.625 0.570 0.681 0.669 0.651 0.644 0.671
#>  [397] 0.562 0.572 0.579 0.699 0.506 0.453 0.656 0.587 0.704 0.577 0.673 0.600
#>  [409] 0.533 0.556 0.593 0.557 0.588 0.701 0.601 0.568 0.574 0.638 0.658 0.640
#>  [421] 0.649 0.633 0.720 0.572 0.658 0.529 0.727 0.558 0.663 0.508 0.704 0.606
#>  [433] 0.663 0.654 0.565 0.694 0.667 0.598 0.559 0.658 0.589 0.577 0.706 0.615
#>  [445] 0.676 0.683 0.560 0.702 0.744 0.587 0.492 0.518 0.548 0.553 0.516 0.526
#>  [457] 0.575 0.756 0.779 0.567 0.595 0.553 0.536 0.585 0.550 0.440 0.600 0.633
#>  [469] 0.678 0.564 0.661 0.642 0.618 0.667 0.678 0.598 0.677 0.598 0.646 0.685
#>  [481] 0.564 0.617 0.515 0.560 0.639 0.639 0.491 0.557 0.580 0.601 0.603 0.664
#>  [493] 0.630 0.678 0.625 0.674 0.628 0.499 0.640 0.606 0.700 0.583 0.719 0.692
#>  [505] 0.658 0.563 0.587 0.615 0.626 0.541 0.706 0.520 0.634 0.551 0.590 0.580
#>  [517] 0.727 0.586 0.537 0.614 0.569 0.693 0.665 0.570 0.629 0.682 0.601 0.548
#>  [529] 0.563 0.610 0.634 0.547 0.673 0.636 0.650 0.478 0.645 0.595 0.666 0.587
#>  [541] 0.638 0.654 0.556 0.643 0.504 0.562 0.556 0.638 0.672 0.531 0.646 0.668
#>  [553] 0.633 0.612 0.565 0.567 0.633 0.667 0.628 0.540 0.663 0.657 0.583 0.563
#>  [565] 0.566 0.717 0.508 0.716 0.733 0.514 0.629 0.544 0.641 0.650 0.655 0.723
#>  [577] 0.572 0.552 0.546 0.579 0.466 0.625 0.488 0.516 0.654 0.653 0.711 0.572
#>  [589] 0.570 0.585 0.669 0.615 0.656 0.653 0.671 0.627 0.671 0.654 0.609 0.612
#>  [601] 0.717 0.700 0.665 0.588 0.604 0.608 0.565 0.546 0.688 0.711 0.656 0.639
#>  [613] 0.561 0.701 0.709 0.761 0.427 0.490 0.684 0.710 0.762 0.570 0.600 0.610
#>  [625] 0.545 0.521 0.597 0.610 0.679 0.601 0.677 0.529 0.570 0.581 0.642 0.713
#>  [637] 0.698 0.460 0.494 0.711 0.604 0.669 0.699 0.631 0.634 0.612 0.697 0.572
#>  [649] 0.650 0.653 0.719 0.549 0.606 0.640 0.561 0.623 0.576 0.675 0.626 0.634
#>  [661] 0.612 0.628 0.577 0.592 0.624 0.700 0.660 0.630 0.581 0.702 0.548 0.604
#>  [673] 0.619 0.672 0.484 0.657 0.624 0.639 0.582 0.585 0.538 0.634 0.669 0.663
#>  [685] 0.677 0.576 0.585 0.558 0.591 0.606 0.608 0.639 0.560 0.565 0.604 0.666
#>  [697] 0.569 0.579 0.620 0.670 0.580 0.563 0.591 0.606 0.709 0.596 0.641 0.650
#>  [709] 0.589 0.608 0.604 0.561 0.638 0.646 0.582 0.618 0.622 0.699 0.618 0.611
#>  [721] 0.646 0.631 0.542 0.467 0.695 0.585 0.588 0.661 0.531 0.688 0.562 0.573
#>  [733] 0.609 0.626 0.616 0.702 0.617 0.574 0.595 0.690 0.570 0.626 0.556 0.509
#>  [745] 0.539 0.557 0.633 0.516 0.554 0.752 0.616 0.610 0.625 0.575 0.748 0.547
#>  [757] 0.650 0.722 0.711 0.493 0.457 0.429 0.795 0.778 0.671 0.689 0.667 0.590
#>  [769] 0.622 0.598 0.640 0.656 0.547 0.451 0.700 0.489 0.536 0.576 0.679 0.630
#>  [781] 0.649 0.685 0.559 0.562 0.640 0.635 0.578 0.669 0.654 0.704 0.619 0.604
#>  [793] 0.576 0.676 0.596 0.673 0.615 0.595 0.591 0.704 0.502 0.620 0.677 0.616
#>  [805] 0.587 0.733 0.707 0.632 0.603 0.619 0.625 0.603 0.647 0.647 0.563 0.662
#>  [817] 0.683 0.593 0.535 0.652 0.661 0.640 0.627 0.638 0.625 0.475 0.490 0.666
#>  [829] 0.545 0.613 0.664 0.624 0.577 0.710 0.793 0.786 0.731 0.502 0.560 0.566
#>  [841] 0.546 0.601 0.662 0.665 0.569 0.588 0.590 0.627 0.699 0.679 0.666 0.617
#>  [853] 0.529 0.580 0.581 0.659 0.608 0.631 0.522 0.511 0.699 0.702 0.702 0.735
#>  [865] 0.613 0.610 0.632 0.727 0.663 0.722 0.675 0.630 0.718 0.589 0.604 0.592
#>  [877] 0.610 0.587 0.668 0.669 0.602 0.675 0.662 0.541 0.579 0.629 0.556 0.638
#>  [889] 0.617 0.666 0.616 0.642 0.628 0.584 0.596 0.587 0.563 0.576 0.640 0.628
#>  [901] 0.718 0.592 0.591 0.589 0.624 0.576 0.715 0.699 0.549 0.493 0.572 0.725
#>  [913] 0.718 0.703 0.690 0.618 0.508 0.715 0.577 0.624 0.586 0.623 0.599 0.613
#>  [925] 0.575 0.689 0.589 0.621 0.600 0.555 0.577 0.597 0.576 0.591 0.578 0.670
#>  [937] 0.609 0.672 0.644 0.624 0.562 0.570 0.563 0.546 0.672 0.578 0.697 0.623
#>  [949] 0.617 0.706 0.520 0.626 0.557 0.659 0.661 0.642 0.624 0.620 0.608 0.631
#>  [961] 0.593 0.528 0.661 0.692 0.673 0.552 0.573 0.634 0.638 0.630 0.620 0.616
#>  [973] 0.645 0.596 0.704 0.574 0.587 0.651 0.638 0.646 0.604 0.678 0.649 0.676
#>  [985] 0.648 0.705 0.704 0.684 0.580 0.600 0.628 0.601 0.617 0.651 0.687 0.629
#>  [997] 0.519 0.580 0.633 0.632 0.545 0.743 0.714 0.655 0.558 0.617 0.588 0.680
#> [1009] 0.523 0.659 0.608 0.602 0.621 0.670 0.552 0.679 0.674 0.649 0.620 0.611
#> [1021] 0.594 0.669 0.528 0.557 0.711 0.683 0.612 0.557 0.663 0.630 0.540 0.557
#> [1033] 0.581 0.649 0.670 0.692 0.563 0.629 0.582 0.611 0.672 0.636 0.685 0.694
#> [1045] 0.683 0.675 0.627 0.614 0.569 0.629 0.645 0.627 0.608 0.571 0.573 0.568
#> [1057] 0.626 0.601 0.567 0.566 0.524 0.507 0.657 0.669 0.670 0.674 0.634 0.709
#> [1069] 0.634 0.583 0.586 0.651 0.644 0.509 0.702 0.627 0.770 0.689 0.581 0.647
#> [1081] 0.626 0.570 0.616 0.659 0.595 0.555 0.626 0.574 0.566 0.614 0.540 0.556
#> [1093] 0.595 0.580 0.666 0.619 0.618 0.644 0.570 0.519 0.476 0.453 0.629 0.594
#> [1105] 0.710 0.672 0.611 0.567 0.589 0.589 0.640 0.582 0.597 0.582 0.535 0.665
#> [1117] 0.676 0.691 0.673 0.656 0.502 0.473 0.439 0.632 0.642 0.584 0.667 0.715
#> [1129] 0.639 0.603 0.560 0.599 0.591 0.745 0.695 0.642 0.616 0.650 0.602 0.658
#> [1141] 0.574 0.542 0.561 0.691 0.584 0.637 0.646 0.655 0.615 0.602 0.720 0.623
#> [1153] 0.658 0.614 0.533 0.654 0.534 0.662 0.610 0.615 0.647 0.653 0.640 0.528
#> [1165] 0.666 0.630 0.623 0.649 0.648 0.612 0.601 0.645 0.720 0.607 0.620 0.607
#> [1177] 0.585 0.681 0.601 0.636 0.570 0.575 0.643 0.560 0.637 0.717 0.662 0.669
#> [1189] 0.614 0.657 0.608 0.577 0.663 0.650 0.717 0.557 0.573 0.554 0.588 0.717
#> [1201] 0.583 0.625 0.586 0.587 0.627 0.618 0.591 0.579 0.614 0.656 0.666 0.604
#> [1213] 0.666 0.618 0.582 0.583 0.672 0.624 0.650 0.618 0.655 0.581 0.599 0.643
#> [1225] 0.623 0.612 0.651 0.585 0.597 0.590 0.559 0.609 0.605 0.704 0.697 0.756
#> [1237] 0.607 0.632 0.546 0.707 0.634 0.592 0.596 0.574 0.674 0.543 0.673 0.619
#> [1249] 0.635 0.588 0.634 0.633 0.689 0.585 0.604 0.560 0.566 0.600 0.567 0.573
#> [1261] 0.565 0.659 0.674 0.658 0.616 0.617 0.521 0.631 0.564 0.611 0.578 0.650
#> [1273] 0.581 0.615 0.530 0.588 0.590 0.583 0.606 0.596 0.611 0.684 0.673 0.724
#> [1285] 0.584 0.501 0.698 0.553 0.553 0.669 0.623 0.540 0.618 0.616 0.618 0.654
#> [1297] 0.641 0.625 0.679 0.642 0.625 0.614 0.614 0.574 0.560 0.640 0.612 0.663
#> [1309] 0.657 0.645 0.601 0.564 0.567 0.707 0.517 0.484 0.631 0.571 0.714 0.566
#> [1321] 0.632 0.655 0.633 0.549 0.661 0.610 0.612 0.593 0.610 0.555 0.521 0.743
#> [1333] 0.658 0.578 0.586 0.637 0.553 0.642 0.680 0.674 0.652 0.677 0.701 0.710
#> [1345] 0.637 0.639 0.663 0.593 0.613 0.545 0.559 0.575 0.564 0.671 0.638 0.572
#> [1357] 0.682 0.682 0.685 0.721 0.684 0.719 0.614 0.666 0.614 0.621 0.537 0.542
#> [1369] 0.481 0.728 0.605 0.560 0.651 0.547 0.618 0.631 0.585 0.655 0.659 0.628
#> [1381] 0.614 0.626 0.515 0.680 0.652 0.627 0.580 0.661 0.651 0.657 0.647 0.480
#> [1393] 0.725 0.616 0.614 0.662 0.659 0.637 0.651 0.659 0.546 0.621 0.602 0.625
#> [1405] 0.602 0.721 0.592 0.526 0.581 0.533 0.563 0.670 0.420 0.810 0.586 0.604
#> [1417] 0.634 0.588 0.641 0.582 0.610 0.642 0.614 0.632 0.596 0.582 0.695 0.665
#> [1429] 0.641 0.619 0.669 0.525 0.592 0.630 0.621 0.628 0.719 0.660 0.565 0.559
#> [1441] 0.533 0.642 0.587 0.578 0.545 0.516 0.645 0.647 0.547 0.690 0.669 0.573
#> [1453] 0.655 0.638 0.746 0.665 0.612 0.669 0.535 0.525 0.666 0.657 0.692 0.661
#> [1465] 0.650 0.608 0.581 0.559 0.714 0.705 0.699 0.574 0.679 0.766 0.512 0.712
#> [1477] 0.672 0.689 0.660 0.665 0.520 0.593 0.587 0.626 0.509 0.612 0.579 0.654
#> [1489] 0.656 0.696 0.580 0.572 0.546 0.523 0.673 0.643 0.661 0.676 0.645 0.608
#> [1501] 0.558 0.559 0.601 0.702 0.614 0.602 0.664 0.566 0.603 0.779 0.500 0.590
#> [1513] 0.577 0.562 0.646 0.657 0.653 0.684 0.630 0.586 0.641 0.624 0.580 0.601
#> [1525] 0.554 0.722 0.653 0.655 0.535 0.625 0.556 0.640 0.648 0.602 0.684 0.652
#> [1537] 0.634 0.636 0.635 0.582 0.603 0.770 0.673 0.527 0.629 0.601 0.614 0.650
#> [1549] 0.509 0.558 0.594 0.645 0.676 0.683 0.626 0.670 0.725 0.761 0.728 0.761
#> [1561] 0.732 0.737 0.690 0.766 0.528 0.543 0.757 0.676 0.652 0.784 0.621 0.649
#> [1573] 0.631 0.566 0.618 0.660 0.554 0.640 0.639 0.541 0.600 0.566 0.537 0.628
#> [1585] 0.628 0.670 0.721 0.516 0.580 0.647 0.752 0.605 0.636 0.513 0.526 0.559
#> [1597] 0.554 0.709 0.608 0.626 0.580 0.708 0.597 0.731 0.600 0.501 0.618 0.711
#> [1609] 0.722 0.679 0.538 0.569 0.607 0.652 0.722 0.527 0.690 0.582 0.628 0.630
#> [1621] 0.664 0.665 0.656 0.617 0.568 0.673 0.657 0.692 0.678 0.710 0.571 0.600
#> [1633] 0.611 0.601 0.554 0.671 0.528 0.667 0.600 0.695 0.589 0.645 0.592 0.619
#> [1645] 0.539 0.667 0.701 0.714 0.524 0.683 0.605 0.568 0.580 0.685 0.604 0.696
#> [1657] 0.585 0.619 0.676 0.753 0.689 0.580 0.654 0.696 0.707 0.568 0.511 0.682
#> [1669] 0.605 0.610 0.486 0.709 0.561 0.717 0.574 0.594 0.617 0.559 0.616 0.709
#> [1681] 0.522 0.505 0.570 0.709 0.675 0.597 0.734 0.675 0.702 0.719 0.635 0.533
#> [1693] 0.543 0.571 0.688 0.682 0.668 0.687 0.600 0.678 0.655 0.644 0.699 0.658
#> [1705] 0.559 0.540 0.520 0.531 0.729 0.717 0.650 0.626 0.692 0.636 0.629 0.636
#> [1717] 0.611 0.642 0.541 0.566 0.598 0.640 0.573 0.586 0.656 0.522 0.580 0.754
#> [1729] 0.715 0.548 0.637 0.602 0.664 0.688 0.741 0.681 0.590 0.616 0.679 0.595
#> [1741] 0.601 0.603 0.668 0.584 0.566 0.569 0.649 0.717 0.732 0.633 0.620 0.603
#> [1753] 0.621 0.552 0.581 0.684 0.628 0.539 0.678 0.672 0.715 0.719 0.708 0.536
#> [1765] 0.645 0.667 0.609 0.604 0.610 0.569 0.524 0.667 0.691 0.714 0.473 0.733
#> [1777] 0.735 0.452 0.665 0.545 0.494 0.591 0.537 0.629 0.660 0.529 0.561 0.612
#> [1789] 0.681 0.701 0.568 0.634 0.623 0.675 0.683 0.521 0.611 0.601 0.610 0.534
#> [1801] 0.664 0.604 0.577 0.552 0.590 0.633 0.600 0.644 0.801 0.455 0.715 0.586
#> [1813] 0.605 0.627 0.607 0.615 0.649 0.673 0.548 0.660 0.676 0.643 0.644 0.620
#> [1825] 0.649 0.579 0.555 0.712 0.565 0.700 0.670 0.563 0.641 0.665 0.640 0.658
#> [1837] 0.707 0.767 0.729 0.522 0.605 0.685 0.577 0.587 0.660 0.631 0.620 0.616
#> [1849] 0.657 0.516 0.597 0.685 0.580 0.575 0.586 0.570 0.532 0.619 0.631 0.635
#> [1861] 0.660 0.675 0.731 0.664 0.563 0.640 0.604 0.684 0.592 0.635 0.666 0.572
#> [1873] 0.611 0.615 0.685 0.647 0.610 0.490 0.680 0.689 0.645 0.574 0.667 0.655
#> [1885] 0.616 0.627 0.574 0.669 0.647 0.590 0.607 0.632 0.583 0.677 0.610 0.616
#> [1897] 0.562 0.694 0.783 0.746 0.728 0.520 0.646 0.484 0.737 0.678 0.598 0.642
#> [1909] 0.688 0.510 0.583 0.613 0.639 0.636 0.556 0.543 0.598 0.584 0.637 0.654
#> [1921] 0.629 0.546 0.534 0.517 0.454 0.505 0.659 0.740 0.583 0.685 0.595 0.658
#> [1933] 0.584 0.677 0.679 0.610 0.589 0.661 0.632 0.617 0.633 0.661 0.668 0.697
#> [1945] 0.693 0.620 0.585 0.665 0.647 0.660 0.643 0.538 0.540 0.549 0.713 0.568
#> [1957] 0.594 0.525 0.540 0.732 0.707 0.732 0.556 0.567 0.709 0.669 0.667 0.605
#> [1969] 0.572 0.613 0.634 0.612 0.697 0.537 0.581 0.659 0.694 0.661 0.669 0.524
#> [1981] 0.510 0.702 0.570 0.650 0.559 0.562 0.642 0.656 0.579 0.722 0.650 0.547
#> [1993] 0.518 0.502 0.770 0.769 0.427 0.369 0.410 0.638
#> 
#> 
#> $model
#> $model$fit
#> stan_glm
#>  family:       binomial [logit]
#>  formula:      y ~ trt:X1
#>  observations: 100
#>  predictors:   3
#> ------
#>             Median MAD_SD
#> (Intercept)  0.7    0.3  
#> trtA:X1     -0.2    0.3  
#> trtC:X1     -0.4    0.2  
#> 
#> ------
#> * For help interpreting the printed output see ?print.stanreg
#> * For info on the priors used see ?prior_summary.stanreg
#> 
#> $model$rho
#> [1] NA
#> 
#> $model$N
#> [1] 1000
#> 
#> $model$stan_args
#> $model$stan_args$algorithm
#> [1] "sampling"
#> 
#> $model$stan_args$chains
#> [1] 2
#> 
#> $model$stan_args$iter
#> [1] 2000
#> 
#> $model$stan_args$refresh
#> [1] 0
#> 
#> $model$stan_args$seed
#> [1] 30956829
#> 
#> 
#>