Calculate the difference between treatments using all evidence

This is the main, top-level wrapper for {outstandR}. Methods taken from RemiroAzocar2022outstandR.

Usage

outstandR(
  ipd_trial,
  ald_trial,
  strategy,
  ref_trt = "C",
  CI = 0.95,
  scale = NULL,
  ...
)

Arguments

ipd_trial

Individual-level patient data. For example, suppose between studies A and C. In a long format and must contain a treatment column and outcome column consistent with the formula object. The labels in the treatment are used internally so there must be a common treatment with the aggregate-level data trial.

ald_trial

Aggregate-level data. For example, suppose between studies B and C. The column names take the form

mean.X*: mean patient measurement

sd.X*: standard deviation of patient measurement

y.*.sum: total number of events

y.*.bar: proportion of events

N.*: total number of individuals

strategy

Computation strategy function. These can be strategy_maic(), strategy_stc(), strategy_gcomp_ml() and strategy_gcomp_stan()

ref_trt

Reference / common / anchoring treatment name; default "C"

CI

Confidence interval; between 0,1

scale

Relative treatment effect scale. If NULL, the scale is automatically determined from the model.

...

Additional arguments

Value

List of length 3 of statistics as a outstandR class object. Containing statistics between each pair of treatments. These are the mean, variances and confidence intervals, for contrasts and absolute values.

References

RemiroAzocar2022outstandR

Examples