Calculate the difference between treatments using all evidence

This is the main, top-level wrapper for {outstandR}. Methods taken from (Remiro‐Azócar et al. 2022) .

Usage

outstandR(
  ipd_trial,
  ald_trial,
  strategy,
  ref_trt = NA,
  CI = 0.95,
  scale = NULL,
  ...
)

Arguments

ipd_trial

Individual-level patient data. For example, suppose between studies A and C. In a long format and must contain a treatment column and outcome column consistent with the formula object. The labels in the treatment are used internally so there must be a common treatment with the aggregate-level data trial.

ald_trial

Aggregate-level data. For example, suppose between studies B and C. The column names are

variable: Covariate name. In the case of treatment arm sample size this is NA

statistic: Summary statistic name from "mean", standard deviation "sd" or "sum"

value: Numerical value of summary statistic

trt: Treatment label. Because we assume a common covariate distribution between treatment arms this is NA

strategy

Computation strategy function. These can be strategy_maic(), strategy_stc(), strategy_gcomp_ml() and strategy_gcomp_stan().

ref_trt

Reference / common / anchoring treatment name; default "C"

CI

Confidence interval; between 0,1

scale

Relative treatment effect scale. If NULL, the scale is automatically determined from the model. Choose from "log-odds", "log_relative_risk", "risk_difference", "delta_z", "mean_difference", "rate_difference" depending on the data type.

...

Additional arguments. Currently, can pass named arguments to rstanarm::stan_glm() via strategy_gcomp_stan().

Value

List of length 3 of statistics as a outstandR class object. Containing statistics between each pair of treatments. These are the mean, variances and confidence intervals, for contrasts and absolute values.

References

Remiro‐Azócar A, Heath A, Baio G (2022). “Parametric G‐computation for compatible indirect treatment comparisons with limited individual patient data.” Res. Synth. Methods, 1–31. ISSN 1759-2879, doi:10.1002/jrsm.1565 , 2108.12208.

See also

Examples